Move utilities dir into lib
