+++ lib/commands/deploy.js
@@ -1,4 +1,4 @@
-var ConfigurationReader = require('../../utilities/configuration-reader');
+var ConfigurationReader = require('../utilities/configuration-reader');
 
 module.exports = {
   name: 'deploy',
