+++ lib/tasks/activate.js
@@ -1,7 +1,7 @@
 var chalk               = require('chalk');
 var Task                = require('ember-cli/lib/models/task');
-var ConfigurationReader = require('../../utilities/configuration-reader');
-var AdapterRegistry     = require('../../utilities/adapter-registry');
+var ConfigurationReader = require('../utilities/configuration-reader');
+var AdapterRegistry     = require('../utilities/adapter-registry');
 var Promise             = require('ember-cli/lib/ext/promise');
 var SilentError         = require('ember-cli/lib/errors/silent');
