+++ lib/tasks/assets.js
@@ -1,5 +1,5 @@
 var Task                = require('ember-cli/lib/models/task');
-var ConfigurationReader = require('../../utilities/configuration-reader');
+var ConfigurationReader = require('../utilities/configuration-reader');
 var Promise             = require('ember-cli/lib/ext/promise');
 var SilentError         = require('ember-cli/lib/errors/silent');
 var chalk               = require('chalk');
@@ -10,7 +10,7 @@ var ncp   = Promise.denodeify(require('ncp'));
 var white = chalk.white;
 var green = chalk.green;
 
-var AssetsUploader = require('../../utilities/assets-uploader');
+var AssetsUploader = require('../utilities/assets-uploader');
 
 var EXPIRE_IN_2030 = new Date('2030');
 var TWO_YEAR_CACHE_PERIOD_IN_SEC = 60 * 60 * 24 * 365 * 2;
