+++ lib/tasks/deploy-index.js
@@ -1,5 +1,5 @@
-var ConfigurationReader = require('../../utilities/configuration-reader');
-var AdapterRegistry     = require('../../utilities/adapter-registry');
+var ConfigurationReader = require('../utilities/configuration-reader');
+var AdapterRegistry     = require('../utilities/adapter-registry');
 var Promise             = require('ember-cli/lib/ext/promise');
 var Task                = require('ember-cli/lib/models/task');
 var chalk               = require('chalk');
