+++ lib/utilities/adapter-registry.js
