+++ lib/utilities/adapter.js
