+++ lib/utilities/assets-uploader.js
