+++ lib/utilities/configuration-reader.js
@@ -1,6 +1,6 @@
 var path      = require('path');
-var Config    = require('../lib/models/config');
 var deprecate = require('./deprecate');
+var Config    = require('../models/config');
 
 function ConfigurationReader(options) {
   var root              = process.cwd();
