+++ lib/utilities/deprecate.js
