+++ lib/utilities/tagging/sha.js
